+++ src/main/kotlin/com/knu/mockin/controller/AccountController.kt
@@ -3,8 +3,6 @@ package com.knu.mockin.controller
 
 import com.knu.mockin.model.dto.request.account.KeyPairRequestDto
 import com.knu.mockin.model.dto.request.account.UserAccountNumberRequestDto
-import com.knu.mockin.model.dto.response.AccessTokenAPIResponseDto
-import com.knu.mockin.model.dto.response.ApprovalKeyResponseDto
 import com.knu.mockin.model.dto.response.SimpleMessageResponseDto
 import com.knu.mockin.service.AccountService
 import org.springframework.http.ResponseEntity
@@ -41,42 +39,6 @@ class AccountController(
         authentication: Authentication
     ): ResponseEntity<SimpleMessageResponseDto> {
         val result = accountService.postRealKeyPair(keyPairRequestDto, authentication.name)
-        return ResponseEntity.ok(result)
-    }
-
-    @PostMapping("/mock-approval-key")
-    suspend fun getMockApprovalKey(
-        authentication: Authentication
-    ): ResponseEntity<ApprovalKeyResponseDto> {
-        val result = accountService.getMockApprovalKey(authentication.name)
-
-        return ResponseEntity.ok(result)
-    }
-
-    @PostMapping("/real-approval-key")
-    suspend fun getRealApprovalKey(
-        authentication: Authentication
-    ): ResponseEntity<ApprovalKeyResponseDto> {
-        val result = accountService.getRealApprovalKey(authentication.name)
-
-        return ResponseEntity.ok(result)
-    }
-
-    @PostMapping("/mock-token")
-    suspend fun getMockAccessToken(
-        authentication: Authentication
-    ): ResponseEntity<AccessTokenAPIResponseDto> {
-        val result = accountService.getMockAccessToken(authentication.name)
-
-        return ResponseEntity.ok(result)
-    }
-
-    @PostMapping("/real-token")
-    suspend fun getRealAccessToken(
-        authentication: Authentication
-    ): ResponseEntity<AccessTokenAPIResponseDto> {
-        val result = accountService.getRealAccessToken(authentication.name)
-
         return ResponseEntity.ok(result)
     }
 }
